♻️ refactor(config): extract `JpaEventStoreConfiguration` and `AxonServerEventStoreConfiguration` from `GameConfiguration`
`GameConfiguration` now only holds the correlation data providers. The event store
wiring lives in two dedicated, mutually exclusive configurations in
`shared/infrastructure`:

- `JpaEventStoreConfiguration` (renamed from `EventStoreConfiguration`,
  active when `axon.axonserver.enabled=false`) — keeps the snapshot-capable
  JPA engine and absorbs the JPA-path `SnapshotStore` bean from
  `GameConfiguration`.
- `AxonServerEventStoreConfiguration` (promoted from a nested class in
  `GameConfiguration`, active when `axon.axonserver.enabled=true`) — the
  aggregate-based engine override and its in-memory `SnapshotStore`.
